BlogPostClient → BlogHeader
